+++ src/index.ts
@@ -1,5 +1,6 @@
 import { Classes as BlueprintClasses } from "@blueprintjs/core";
 
+import {helperClasses as reactFlowHelperClasses } from "./cmem/react-flow/ReactFlow/constants";
 import { ClassNames as IntentClassNames } from "./common/Intent";
 import * as Skeleton from "./components/Skeleton/classnames";
 import * as TypographyClassNames from "./components/Typography/classnames";
@@ -9,7 +10,11 @@ const ClassNames = {
     Intent: IntentClassNames,
     Skeleton,
     Typography: TypographyClassNames,
+    ReactFlow: reactFlowHelperClasses,
 };
+
+/** @deprecated (v27) use `ClassNames.ReactFlow.preventAllActions` */
+const preventReactFlowActionsClasses = reactFlowHelperClasses.preventAllActions;
 
 export * from "./configuration/constants";
 export * from "./common";
@@ -17,4 +22,4 @@ export * from "./components";
 export * from "./extensions";
 export * from "./cmem";
 
-export { ClassNames };
+export { ClassNames, preventReactFlowActionsClasses };